# perf: apply transformer and const enum optimize