--- a/src/container/ContainerBase/index.ts
+++ b/src/container/ContainerBase/index.ts
@@ -1,12 +1,15 @@
+export const enum IteratorType {
+  NORMAL = 0,
+  REVERSE = 1
+}
+
 export abstract class ContainerIterator<T> {
-  static readonly NORMAL = false;
-  static readonly REVERSE = true;
   /**
    * @description Iterator's type.
    */
-  readonly iteratorType: boolean;
+  readonly iteratorType: IteratorType;
   protected node: unknown;
-  protected constructor(iteratorType: boolean = ContainerIterator.NORMAL) {
+  protected constructor(iteratorType: IteratorType = IteratorType.NORMAL) {
     this.iteratorType = iteratorType;
   }
   /**