--- a/src/container/TreeContainer/Base/TreeIterator.ts
+++ b/src/container/TreeContainer/Base/TreeIterator.ts
@@ -1,5 +1,5 @@
 import TreeNode from './TreeNode';
-import { ContainerIterator } from '@/container/ContainerBase/index';
+import { ContainerIterator, IteratorType } from '@/container/ContainerBase/index';
 
 abstract class TreeIterator<K, V> extends ContainerIterator<K | [K, V]> {
   protected node: TreeNode<K, V>;
@@ -9,13 +9,13 @@ abstract class TreeIterator<K, V> extends ContainerIterator<K | [K, V]> {
   constructor(
     node: TreeNode<K, V>,
     header: TreeNode<K, V>,
-    iteratorType?: boolean
+    iteratorType?: IteratorType
   ) {
     super(iteratorType);
     this.node = node;
     this.header = header;
 
-    if (this.iteratorType === ContainerIterator.NORMAL) {
+    if (this.iteratorType === IteratorType.NORMAL) {
       this.pre = function () {
         if (this.node === this.header.left) {
           throw new RangeError('LinkList iterator access denied!');