<!--
Please make sure to read the Pull Request Guidelines:
https://github.com/js-sdsl/js-sdsl/blob/main/.github/CONTRIBUTING.md
-->

<!-- PULL REQUEST TEMPLATE -->
<!-- (Update "[ ]" to "[x]" to check a box) -->

**What kind of change does this PR introduce?** (check at least one)

- [ ] Bugfix
- [ ] Feature
- [x] Code style update
- [x] Refactor
- [x] Build-related changes
- [ ] Other, please describe:

**Does this PR introduce a breaking change?** (check one)

- [x] Yes
- [ ] No

If yes, please describe the impact and migration path for existing applications:

Migration is not required. The revised contents are as follows

- The parameters of the constructor of the `ContainerIterator` type have been changed. This is a breaking change, but considering the structure of the library, the `ContainerIterator` type seems like not considered to be inherited from outside the library, so it is judged not to be a problem.
- The type of color property of `TreeNode` has been changed But this also doesn't seem to be a problem for the above reasons.
- The constants in the `HashContainer` have been moved to const enum However, I don't think this will be a problem for the above reasons.
- The `checkWithinAccessParams` function is no longer a public API. It was a necessary task in the process of switching to a macro for optimization. But this function also seems fine because it doesn't seem to have been designed to be used outside of the library.

**The PR fulfills these requirements:**

- [x] It's submitted to the `dev` branch, **not** the `main` branch
- [x] When resolving a specific issue, it's referenced in the PR's title (e.g. `fix #xxx[,#xxx]`, where "xxx" is the issue number)
- [x] All tests are passing
- [x] New/updated tests are included

If adding a **new feature**, the PR's description includes:
- [x] A convincing reason for adding this feature (to avoid wasting your time, it's best to open a suggestion issue first and wait for approval before working on it)

# What I Changed
- applied [gulp](https://gulpjs.com/) to build workflow.
- applied [ts-macros](https://github.com/GoogleFeud/ts-macros) and [ts-transformer-minify-privates](https://github.com/timocov/ts-transformer-minify-privates) transformers.
- property mangling was performed using terser in the build.
- constants are moved to const enums

## Private property mangling
Introducing gulp into build workflows was necessary for typescript transformers for optimization.

The gulp build performs a typescript build and then mangle the private property through terser.

Property mangling is not usually used because it destroys code. However, if ts-transformer-minify-privates and terser are used together, it is possible to do this safely.

### Expectation effectiveness
- The bundle size is reduced because the length of the private property is short to about 2 characters
- Property contractions can be expected to improve performance. The rationale is [this](https://stackoverflow.com/questions/31719887/variable-name-length-vs-performance).

I confirmed that the umd.min **build size has been reduced by 0.4KB**. The more private properties you increase, the greater this difference will be.
![mangle result](https://user-images.githubusercontent.com/48761044/189035278-0070199d-17ae-45c6-ace0-bde4e219ce75.png)


## Change `checkWithinAccessParams` function into macro

To use macros, ts-macros typescript transformer was applied to the build.

I checked that checkWithinAccessParams is moderately short and frequently called, and I decided that it would be wise to change it to macro.

#### Expectation effectiveness
Utility functions that determine the range are frequently invoked. Turning these moderately short functions into **macros eliminates the function call cost and improves performance**.

![macro result](https://user-images.githubusercontent.com/48761044/189035986-addc774b-7f2d-47b7-8311-1e1288a3c1c5.png)
Build results. You can see that the call to the function disappears and the body of the function is substituted.

## Const enums

I checked that the constants in the `ContainerIterator`, `HashContainer`, and `TreeNode` types are frequently used. This has been replaced with a const enum.

### Expectation effectiveness
The value in the const enum is evaluated at compile time, and the code for getting the value is replaced with that value.

![const enum result](https://user-images.githubusercontent.com/48761044/189036901-bea0d354-5c36-46fb-ad14-691203b1f3bb.png)
Build results. The cost of reading constants is all gone.

# Banchmark

It has not yet been accurately benchmarked. I'll add it as soon as possible.

I believe that this commit will bring real performance improvements.