@noname0310 2022-09-08
I just checked the reason for the reduced code coverage. The checkWithinAccessParams function was changed to macro, and a case that was not previously tracked was found.

Regarding this, I think it will be difficult for me to write the test code because it is not the code I wrote.