@ZLY201 2022-09-08
Are you sure this is caused by macro? Emm.. it doesn't look like. It's also possible that I don't know much about macro.

I'll go to study macro and gulp and evaluate them.

This PR will be pending during this time.

Thanks again.