@noname0310 2022-09-08
Looking closely now, I don't think it's an issue about macros.
It just seems that there is no code to test the method.
I'm writing the test code for now.