@noname0310 2022-09-08
Because the test code uses Math.random, there is a problem that the test cannot pass probably...