@noname0310 2022-09-08
```
$ yarn test:unit && yarn test:performance
$ jest --coverage
 PASS  test/SequentialContainerTest/Deque.test.ts (32.968 s)                                                                                                                                                                                                     
 PASS  test/SequentialContainerTest/LinkList.test.ts (36.789 s)
 PASS  test/OtherTest/iterator.test.ts (42.459 s)
 PASS  test/OtherContainerTest/PriorityQueue.test.ts
 PASS  test/OtherContainerTest/Stack.test.ts
 PASS  test/OtherContainerTest/Queue.test.ts
 PASS  test/SequentialContainerTest/SequentialContainer.test.ts (12.659 s)
 PASS  test/OtherTest/constructor.test.ts
 PASS  test/OtherTest/symbol-iterator.test.ts (10.87 s)
 PASS  test/HashContainerTest/HashSet.test.ts (62.555 s)
 PASS  test/TreeContainerTest/OrderedMap.test.ts (81.317 s)
 PASS  test/TreeContainerTest/OrderedSet.test.ts (91.734 s)
 PASS  test/HashContainerTest/HashMap.test.ts (104.36 s)
----------------------------------------|---------|----------|---------|---------|-------------------
File                                    | % Stmts | % Branch | % Funcs | % Lines | Uncovered Line #s 
----------------------------------------|---------|----------|---------|---------|-------------------
All files                               |   99.54 |    98.49 |     100 |   99.56 |                   
 src                                    |     100 |      100 |     100 |     100 |                   
  index.ts                              |     100 |      100 |     100 |     100 |                   
 src/container/ContainerBase            |     100 |      100 |     100 |     100 |                   
  index.ts                              |     100 |      100 |     100 |     100 |                   
 src/container/HashContainer            |   97.04 |    90.54 |     100 |    97.2 | 
  HashMap.ts                            |   97.65 |     92.5 |     100 |   97.41 | 49,60,154        
  HashSet.ts                            |   96.33 |    88.23 |     100 |   96.96 | 50,61,127        
 src/container/HashContainer/Base       |     100 |      100 |     100 |     100 | 
  index.ts                              |     100 |      100 |     100 |     100 | 
 src/container/OtherContainer           |     100 |      100 |     100 |     100 | 
  PriorityQueue.ts                      |     100 |      100 |     100 |     100 | 
  Queue.ts                              |     100 |      100 |     100 |     100 | 
  Stack.ts                              |     100 |      100 |     100 |     100 | 
 src/container/SequentialContainer      |     100 |      100 |     100 |     100 | 
  Deque.ts                              |     100 |      100 |     100 |     100 | 
  LinkList.ts                           |     100 |      100 |     100 |     100 | 
  Vector.ts                             |     100 |      100 |     100 |     100 | 
 src/container/SequentialContainer/Base |     100 |      100 |     100 |     100 | 
  RandomIterator.ts                     |     100 |      100 |     100 |     100 | 
  index.ts                              |     100 |      100 |     100 |     100 | 
 src/container/TreeContainer            |     100 |      100 |     100 |     100 | 
  OrderedMap.ts                         |     100 |      100 |     100 |     100 | 
  OrderedSet.ts                         |     100 |      100 |     100 |     100 | 
 src/container/TreeContainer/Base       |     100 |      100 |     100 |     100 | 
  TreeIterator.ts                       |     100 |      100 |     100 |     100 | 
  TreeNode.ts                           |     100 |      100 |     100 |     100 | 
  index.ts                              |     100 |      100 |     100 |     100 | 
 src/utils                              |     100 |      100 |     100 |     100 | 
  Mulberry32.ts                         |     100 |      100 |     100 |     100 | 
  checkParams.ts                        |     100 |      100 |     100 |     100 |                   
----------------------------------------|---------|----------|---------|---------|-------------------

Test Suites: 13 passed, 13 total
Tests:       104 passed, 104 total
Snapshots:   0 total
Time:        106.664 s
```
I filled out the coverage by writing the fuzz test now, but those three lines are not passed.
Is it a line that can't be approached logically?