@noname0310 2022-09-10
# here is benchmark results

test environment:

CPU: Intel® Core™ i7-8565U Processor (clock has fixed to 2.6GHz by [ThrottleStop](https://www.techpowerup.com/download/techpowerup-throttlestop/))
RAM: 24GB DDR4
node version: v16.17.0

## `main` branch

```
C:\Users\noname\Desktop\js-sdsl-benchmark\main\performance>node --expose-gc --max_old_space_size=16384 index.js
deque get element by pos with 100 iterations: avg: 5578.71ms, min: 5439ms, max: 10303ms
deque set element by pos with 100 iterations: avg: 8705.69ms, min: 8647ms, max: 8874ms
orderedset insert and delete with 100 iterations: avg: 33160.92ms, min: 32853ms, max: 37918ms
hashset insert and delete with 100 iterations: avg: 5649.9ms, min: 5599ms, max: 5836ms
```

## `perf/tramsformer` branch

```
C:\Users\noname\Desktop\js-sdsl-benchmark\transformer\performance>node --expose-gc --max_old_space_size=16384 index.js
deque get element by pos with 100 iterations: avg: 5512.52ms, min: 5428ms, max: 10128ms
deque set element by pos with 100 iterations: avg: 8652.6ms, min: 8596ms, max: 8825ms
orderedset insert and delete with 100 iterations: avg: 33024.28ms, min: 32719ms, max: 37669ms
hashset insert and delete with 100 iterations: avg: 5642.11ms, min: 5588ms, max: 5852ms
```

## summary
- It's not a surprising improvement because there's no change in the algorithm, but it's definitely faster than before.
- For accurate testing, the time was measured by averaging the tests 100 times.
- Since the test takes too much time, i only wrote the test based on the changed codes. (It took 5 hours to run this test.)

test code:
```typescript
import { Deque, OrderedSet } from "@/index";

class Mulberry32 {
  private seed: number;
  constructor(seed: number) {
    this.seed = seed;
  }

  next(): number {
    let t = this.seed += 0x6D2B79F5;
    t = Math.imul(t ^ t >>> 15, t | 1);
    t ^= t + Math.imul(t ^ t >>> 7, t | 61);
    return ((t ^ t >>> 14) >>> 0) / 4294967296;
  }
}

function runTest(testName: string, test: () => void, iterations: number) {
  let avg = 0;
  let min = Number.MAX_VALUE;
  let max = Number.MIN_VALUE;
  for (let i = 0; i < iterations; i++) {
    global.gc?.();
    let start = Date.now();
    test();
    let end = Date.now();
    let time = end - start;
    avg += time;
    min = Math.min(min, time);
    max = Math.max(max, time);
  }
  avg /= iterations;
  console.log(`${testName} with ${iterations} iterations: avg: ${avg}ms, min: ${min}ms, max: ${max}ms`);
}

function dequeGetElementByPosTest(testNum: number) {
  const mulberry = new Mulberry32(123456789);
  const deque = new Deque<number>();
  for (let i = 0; i < testNum; i++) {
    deque.pushBack(mulberry.next());
  }

  for (let i = 0; i < testNum; i++) {
    deque.getElementByPos(Math.floor(mulberry.next() * testNum));
  }
}

function dequeSetElementByPosTest(testNum: number) {
  const mulberry = new Mulberry32(123456789);
  const deque = new Deque<number>();
  for (let i = 0; i < testNum; i++) {
    deque.pushBack(mulberry.next());
  }

  for (let i = 0; i < testNum; i++) {
    deque.setElementByPos(Math.floor(mulberry.next() * testNum), mulberry.next());
  }
}

function orderedSetInsertAndDeleteTest(testNum: number) {
  const mulberry = new Mulberry32(123456789);
  const set = new OrderedSet<number>();
  for (let i = 0; i < testNum; i++) {
    set.insert(Math.floor(mulberry.next() * testNum));
  }

  for (let i = 0; i < testNum; i++) {
    if (mulberry.next() < 0.5) {
      set.eraseElementByKey(Math.floor(mulberry.next() * testNum));
    } else {
      set.insert(Math.floor(mulberry.next() * testNum));
    }
  }
}

function hashSetInsertAndDeleteTest(testNum: number) {
  const mulberry = new Mulberry32(123456789);
  const set = new Set<number>();
  for (let i = 0; i < testNum; i++) {
    set.add(Math.floor(mulberry.next() * testNum));
  }

  for (let i = 0; i < testNum; i++) {
    if (mulberry.next() < 0.5) {
      set.delete(Math.floor(mulberry.next() * testNum));
    } else {
      set.add(Math.floor(mulberry.next() * testNum));
    }
  }
}

function main() {
  const iterations = 100;
  runTest("deque get element by pos", () => dequeGetElementByPosTest(100000000), iterations);
  runTest("deque set element by pos", () => dequeSetElementByPosTest(100000000), iterations);
  runTest("orderedset insert and delete", () => orderedSetInsertAndDeleteTest(10000000), iterations);
  runTest("hashset insert and delete", () => hashSetInsertAndDeleteTest(10000000), iterations);
}

main();
```